Cinder kilo-3 milestone (2015.1.0b3)

+++ cinder/tests/test_emc_vmax.py
@@ -138,6 +138,11 @@ class Fake_CIMProperty():
         cimproperty.value = 'OS-myhost-MV'
         return cimproperty
 
+    def fake_getSupportedReplicationTypes(self):
+        cimproperty = Fake_CIMProperty()
+        cimproperty.value = [2L, 10L]
+        return cimproperty
+
 
 class Fake_CIM_TierPolicyServiceCapabilities():
 
@@ -540,6 +545,8 @@ class FakeEcomConnection():
             result = self._enum_storagesystems()
         elif name == 'Symm_TierPolicyRule':
             result = self._enum_policyrules()
+        elif name == 'CIM_ReplicationServiceCapabilities':
+            result = self._enum_repservcpbls()
         else:
             result = self._default_enum()
         return result
@@ -590,6 +597,8 @@ class FakeEcomConnection():
             result = self._getinstance_storagehardwareid(objectpath)
         elif name == 'Symm_VirtualProvisioningPool':
             result = self._getinstance_pool(objectpath)
+        elif name == 'Symm_ReplicationServiceCapabilities':
+            result = self._getinstance_replicationServCapabilities(objectpath)
         else:
             result = self._default_getinstance(objectpath)
 
@@ -672,6 +681,8 @@ class FakeEcomConnection():
             result = self._enum_initMaskingGroup()
         elif ResultClass == 'Symm_LunMaskingView':
             result = self._enum_maskingView()
+        elif ResultClass == 'EMC_Meta':
+            result = self._enum_metavolume()
         else:
             result = self._default_assocnames(objectpath)
         return result
@@ -1012,6 +1023,15 @@ class FakeEcomConnection():
         svInstance['SystemElement'] = 'SystemElement'
         svInstance['PercentSynced'] = 100
         return svInstance
+
+    def _getinstance_replicationServCapabilities(self, objectpath):
+        repServCpblInstance = SYMM_SrpStoragePool()
+        classcimproperty = Fake_CIMProperty()
+        repTypesCimproperty = (
+            classcimproperty.fake_getSupportedReplicationTypes())
+        properties = {u'SupportedReplicationTypes': repTypesCimproperty}
+        repServCpblInstance.properties = properties
+        return repServCpblInstance
 
     def _default_getinstance(self, objectpath):
         return objectpath
@@ -1377,6 +1397,9 @@ class FakeEcomConnection():
         portgroups.append(portgroup)
         return portgroups
 
+    def _enum_metavolume(self):
+        return []
+
     def _default_enum(self):
         names = []
         name = {}
@@ -1447,6 +1470,34 @@ class EMCVMAXISCSIDriverNoFastTestCase(test.TestCase):
         doc.writexml(f)
         f.close()
 
+    def create_fake_config_file_no_fast_with_interval(self):
+
+        doc = minidom.Document()
+        emc = doc.createElement("EMC")
+        doc.appendChild(emc)
+        doc = self.add_array_info(doc, emc)
+        doc = self.add_interval_only(doc, emc)
+        filename = 'cinder_emc_config_ISCSINoFAST.xml'
+        self.config_file_path = self.tempdir + '/' + filename
+
+        f = open(self.config_file_path, 'w')
+        doc.writexml(f)
+        f.close()
+
+    def create_fake_config_file_no_fast_with_retries(self):
+
+        doc = minidom.Document()
+        emc = doc.createElement("EMC")
+        doc.appendChild(emc)
+        doc = self.add_array_info(doc, emc)
+        doc = self.add_retries_only(doc, emc)
+        filename = 'cinder_emc_config_ISCSINoFAST.xml'
+        self.config_file_path = self.tempdir + '/' + filename
+
+        f = open(self.config_file_path, 'w')
+        doc.writexml(f)
+        f.close()
+
     def add_array_info(self, doc, emc):
         array = doc.createElement("Array")
         arraytext = doc.createTextNode("1234567891011")
@@ -1505,6 +1556,20 @@ class EMCVMAXISCSIDriverNoFastTestCase(test.TestCase):
 
         retries = doc.createElement("Retries")
         retriestext = doc.createTextNode("40")
+        emc.appendChild(retries)
+        retries.appendChild(retriestext)
+        return doc
+
+    def add_interval_only(self, doc, emc):
+        interval = doc.createElement("Interval")
+        intervaltext = doc.createTextNode("20")
+        emc.appendChild(interval)
+        interval.appendChild(intervaltext)
+        return doc
+
+    def add_retries_only(self, doc, emc):
+        retries = doc.createElement("Retries")
+        retriestext = doc.createTextNode("70")
         emc.appendChild(retries)
         retries.appendChild(retriestext)
         return doc
@@ -1701,7 +1766,10 @@ class EMCVMAXISCSIDriverNoFastTestCase(test.TestCase):
         volumeInstanceName = (
             conn.EnumerateInstanceNames("EMC_StorageVolume")[0])
         volumeName = "1403160-Vol"
-        extraSpecs = self.driver.common.extraSpecs
+        extraSpecs = {'volume_backend_name': 'ISCSINoFAST'}
+        extraSpecs = (
+            self.driver.common._get_job_extra_specs(self.config_file_path,
+                                                    extraSpecs))
 
         # Deleting Storage Group failed
         self.assertRaises(
@@ -2013,34 +2081,50 @@ class EMCVMAXISCSIDriverNoFastTestCase(test.TestCase):
             'OS-PORTGROUP' in
             self.driver.utils.parse_file_to_get_port_group_name(
                 self.config_file_1364232))
-        bExists = os.path.exists(self.config_file_1364232)
-        if bExists:
-            os.remove(self.config_file_1364232)
 
-    @mock.patch.object(
-        emc_vmax_common.EMCVMAXCommon,
-        '_get_pool_and_storage_system',
-        return_value=(None, EMCVMAXCommonData.storage_system))
-    @mock.patch.object(
-        volume_types,
-        'get_volume_type_extra_specs',
-        return_value={'volume_backend_name': 'ISCSINoFAST'})
-    def test_intervals_and_retries(
-            self, _mock_volume_type, mock_storage_system):
-        save_config_path = self.config_file_path
+    def test_intervals_and_retries_override(
+            self):
         self.create_fake_config_file_no_fast_with_add_ons()
-        self.driver.create_volume(self.data.test_volume_v2)
-        extraSpecs = self.driver.common.extraSpecs
+        extraSpecs = {'volume_backend_name': 'ISCSINoFAST'}
+        extraSpecs = (
+            self.driver.common._get_job_extra_specs(self.config_file_path,
+                                                    extraSpecs))
         self.assertEqual(40,
                          self.driver.utils._get_max_job_retries(extraSpecs))
         self.assertEqual(5,
                          self.driver.utils._get_interval_in_secs(extraSpecs))
 
-        bExists = os.path.exists(self.config_file_path)
-        if bExists:
-            os.remove(self.config_file_path)
+    def test_intervals_and_retries_default(self):
+        extraSpecs = {'volume_backend_name': 'ISCSINoFAST'}
+        extraSpecs = (
+            self.driver.common._get_job_extra_specs(self.config_file_path,
+                                                    extraSpecs))
+        self.assertEqual(60,
+                         self.driver.utils._get_max_job_retries(extraSpecs))
+        self.assertEqual(10,
+                         self.driver.utils._get_interval_in_secs(extraSpecs))
 
-        self.config_file_path = save_config_path
+    def test_interval_only(self):
+        extraSpecs = {'volume_backend_name': 'ISCSINoFAST'}
+        self.create_fake_config_file_no_fast_with_interval()
+        extraSpecs = (
+            self.driver.common._get_job_extra_specs(self.config_file_path,
+                                                    extraSpecs))
+        self.assertEqual(60,
+                         self.driver.utils._get_max_job_retries(extraSpecs))
+        self.assertEqual(20,
+                         self.driver.utils._get_interval_in_secs(extraSpecs))
+
+    def test_retries_only(self):
+        extraSpecs = {'volume_backend_name': 'ISCSINoFAST'}
+        self.create_fake_config_file_no_fast_with_retries()
+        extraSpecs = (
+            self.driver.common._get_job_extra_specs(self.config_file_path,
+                                                    extraSpecs))
+        self.assertEqual(70,
+                         self.driver.utils._get_max_job_retries(extraSpecs))
+        self.assertEqual(10,
+                         self.driver.utils._get_interval_in_secs(extraSpecs))
 
     @mock.patch.object(
         emc_vmax_utils.EMCVMAXUtils,
@@ -2545,9 +2629,14 @@ class EMCVMAXISCSIDriverNoFastTestCase(test.TestCase):
             self.data.test_ctxt, self.data.test_CG_snapshot)
 
     def _cleanup(self):
-        bExists = os.path.exists(self.config_file_path)
-        if bExists:
-            os.remove(self.config_file_path)
+        if self.config_file_path:
+            bExists = os.path.exists(self.config_file_path)
+            if bExists:
+                os.remove(self.config_file_path)
+        if self.config_file_1364232:
+            bExists = os.path.exists(self.config_file_1364232)
+            if bExists:
+                os.remove(self.config_file_1364232)
         shutil.rmtree(self.tempdir)
 
 
@@ -2994,31 +3083,6 @@ class EMCVMAXISCSIDriverFastTestCase(test.TestCase):
                           self.driver.create_volume_from_snapshot,
                           self.data.test_volume,
                           EMCVMAXCommonData.test_source_volume)
-
-    @mock.patch.object(
-        emc_vmax_utils.EMCVMAXUtils,
-        'get_volume_meta_head',
-        return_value=None)
-    @mock.patch.object(
-        emc_vmax_common.EMCVMAXCommon,
-        '_find_storage_sync_sv_sv',
-        return_value=(None, None))
-    @mock.patch.object(
-        FakeDB,
-        'volume_get',
-        return_value=EMCVMAXCommonData.test_source_volume)
-    @mock.patch.object(
-        volume_types,
-        'get_volume_type_extra_specs',
-        return_value={'volume_backend_name': 'ISCSIFAST'})
-    def test_create_clone_simple_volume_fast_success(
-            self, mock_volume_type, mock_volume, mock_sync_sv,
-            mock_simple_volume):
-        self.data.test_volume['volume_name'] = "vmax-1234567"
-        self.driver.common.fast.is_volume_in_default_SG = (
-            mock.Mock(return_value=True))
-        self.driver.create_cloned_volume(self.data.test_volume,
-                                         EMCVMAXCommonData.test_source_volume)
 
     @mock.patch.object(
         emc_vmax_common.EMCVMAXCommon,
@@ -4098,28 +4162,23 @@ class EMCVMAXFCDriverFastTestCase(test.TestCase):
                           self.data.test_volume,
                           EMCVMAXCommonData.test_source_volume)
 
-    @mock.patch.object(
-        emc_vmax_utils.EMCVMAXUtils,
-        'get_volume_meta_head',
-        return_value=None)
-    @mock.patch.object(
-        emc_vmax_common.EMCVMAXCommon,
-        '_find_storage_sync_sv_sv',
-        return_value=(None, None))
-    @mock.patch.object(
-        FakeDB,
-        'volume_get',
-        return_value=EMCVMAXCommonData.test_source_volume)
-    @mock.patch.object(
-        volume_types,
-        'get_volume_type_extra_specs',
-        return_value={'volume_backend_name': 'FCFAST'})
-    def test_create_clone_simple_volume_fast_success(
-            self, mock_volume_type, mock_volume, mock_sync_sv,
-            mock_simple_volume):
+    def test_create_clone_simple_volume_fast_success(self):
+        extraSpecs = {'storagetype:fastpolicy': 'FC_GOLD1',
+                      'volume_backend_name': 'FCFAST',
+                      'isV3': False}
+        self.driver.common._initial_setup = (
+            mock.Mock(return_value=extraSpecs))
+        self.driver.common.extraSpecs = extraSpecs
+        self.driver.utils.is_clone_licensed = (
+            mock.Mock(return_value=True))
+        FakeDB.volume_get = (
+            mock.Mock(return_value=EMCVMAXCommonData.test_source_volume))
         self.data.test_volume['volume_name'] = "vmax-1234567"
         self.driver.common.fast.is_volume_in_default_SG = (
             mock.Mock(return_value=True))
+        self.driver.utils.isArrayV3 = mock.Mock(return_value=False)
+        self.driver.common._find_storage_sync_sv_sv = (
+            mock.Mock(return_value=(None, None)))
         self.driver.create_cloned_volume(self.data.test_volume,
                                          EMCVMAXCommonData.test_source_volume)
 
@@ -4270,6 +4329,19 @@ class EMCVMAXFCDriverFastTestCase(test.TestCase):
             self, _mock_volume_type, _mock_storage):
         self.driver.delete_cgsnapshot(
             self.data.test_ctxt, self.data.test_CG_snapshot)
+
+    # Bug 1385450
+    def test_create_clone_without_license(self):
+        mockRepServCap = {}
+        mockRepServCap['InstanceID'] = 'SYMMETRIX+1385450'
+        self.driver.utils.find_replication_service_capabilities = (
+            mock.Mock(return_value=mockRepServCap))
+        self.driver.utils.is_clone_licensed = (
+            mock.Mock(return_value=False))
+        self.assertRaises(exception.VolumeBackendAPIException,
+                          self.driver.create_cloned_volume,
+                          self.data.test_volume,
+                          EMCVMAXCommonData.test_source_volume)
 
     def _cleanup(self):
         bExists = os.path.exists(self.config_file_path)
